+++ src/main/java/redis/clients/jedis/Pipeline.java
@@ -28,7 +28,7 @@ public class Pipeline implements Commands {
         client.decr(key);
     }
 
-    public void decrBy(String key, int integer) {
+    public void decrBy(String key, long integer) {
         client.decrBy(key, integer);
     }
 
@@ -76,7 +76,7 @@ public class Pipeline implements Commands {
         client.hgetAll(key);
     }
 
-    public void hincrBy(String key, String field, int value) {
+    public void hincrBy(String key, String field, long value) {
         client.hincrBy(key, field, value);
     }
 
@@ -112,7 +112,7 @@ public class Pipeline implements Commands {
         client.incr(key);
     }
 
-    public void incrBy(String key, int integer) {
+    public void incrBy(String key, long integer) {
         client.incrBy(key, integer);
     }
